LPS-75829 Extract duplicate methods to new util class
