+++ modules/sdk/gradle-plugins-js-transpiler/src/main/java/com/liferay/gradle/plugins/js/transpiler/JSTranspilerBasePlugin.java
@@ -14,6 +14,7 @@
 
 package com.liferay.gradle.plugins.js.transpiler;
 
+import com.liferay.gradle.plugins.js.transpiler.internal.util.JSTranspilerPluginUtil;
 import com.liferay.gradle.plugins.node.NodePlugin;
 import com.liferay.gradle.plugins.node.tasks.NpmInstallTask;
 import com.liferay.gradle.util.GradleUtil;
@@ -21,16 +22,11 @@ import com.liferay.gradle.util.copy.RenameDependencyClosure;
 
 import java.io.File;
 
-import java.util.HashSet;
-import java.util.Set;
-
 import org.gradle.api.Action;
 import org.gradle.api.Plugin;
 import org.gradle.api.Project;
 import org.gradle.api.Task;
 import org.gradle.api.artifacts.Configuration;
-import org.gradle.api.artifacts.DependencySet;
-import org.gradle.api.artifacts.ProjectDependency;
 import org.gradle.api.tasks.Copy;
 import org.gradle.api.tasks.TaskDependency;
 
@@ -82,44 +78,6 @@ public class JSTranspilerBasePlugin implements Plugin<Project> {
 		return configuration;
 	}
 
-	private Copy _addTaskExpandCompileDependency(
-		Project project, File file, File destinationDir, String taskNamePrefix,
-		RenameDependencyClosure renameDependencyClosure) {
-
-		String taskName = GradleUtil.getTaskName(taskNamePrefix, file);
-
-		Copy copy = GradleUtil.addTask(project, taskName, Copy.class);
-
-		copy.doFirst(
-			new Action<Task>() {
-
-				@Override
-				public void execute(Task task) {
-					Copy copy = (Copy)task;
-
-					Project project = copy.getProject();
-
-					project.delete(copy.getDestinationDir());
-				}
-
-			});
-
-		copy.from(project.zipTree(file));
-
-		String name = renameDependencyClosure.call(file.getName());
-
-		name = name.substring(0, name.length() - 4);
-
-		destinationDir = new File(destinationDir, name);
-
-		copy.setDescription(
-			"Expands " + file.getName() + " into " +
-				project.relativePath(destinationDir) + ".");
-		copy.setDestinationDir(destinationDir);
-
-		return copy;
-	}
-
 	private Task _addTaskExpandJSCompileDependencies(Project project) {
 		Task task = project.task(EXPAND_JS_COMPILE_DEPENDENCIES_TASK_NAME);
 
@@ -138,11 +96,11 @@ public class JSTranspilerBasePlugin implements Plugin<Project> {
 		RenameDependencyClosure renameDependencyClosure =
 			new RenameDependencyClosure(project, configuration.getName());
 
-		Iterable<TaskDependency> taskDependencies = _getTaskDependencies(
-			configuration);
+		Iterable<TaskDependency> taskDependencies =
+			JSTranspilerPluginUtil.getTaskDependencies(configuration);
 
 		for (File file : configuration) {
-			Copy copy = _addTaskExpandCompileDependency(
+			Copy copy = JSTranspilerPluginUtil.addTaskExpandCompileDependency(
 				project, file, npmInstallTask.getNodeModulesDir(),
 				"expandJSCompileDependency", renameDependencyClosure);
 
@@ -151,22 +109,6 @@ public class JSTranspilerBasePlugin implements Plugin<Project> {
 
 			expandJSCompileDependenciesTask.dependsOn(copy);
 		}
-	}
-
-	private Iterable<TaskDependency> _getTaskDependencies(
-		Configuration configuration) {
-
-		Set<TaskDependency> taskDependencies = new HashSet<>();
-
-		DependencySet dependencySet = configuration.getAllDependencies();
-
-		for (ProjectDependency projectDependency : dependencySet.withType(
-				ProjectDependency.class)) {
-
-			taskDependencies.add(projectDependency.getBuildDependencies());
-		}
-
-		return taskDependencies;
 	}
 
 }
