+++ modules/sdk/gradle-plugins-js-transpiler/src/main/java/com/liferay/gradle/plugins/js/transpiler/JSTranspilerPlugin.java
@@ -14,6 +14,7 @@
 
 package com.liferay.gradle.plugins.js.transpiler;
 
+import com.liferay.gradle.plugins.js.transpiler.internal.util.JSTranspilerPluginUtil;
 import com.liferay.gradle.plugins.node.NodePlugin;
 import com.liferay.gradle.plugins.node.tasks.DownloadNodeModuleTask;
 import com.liferay.gradle.plugins.node.tasks.ExecuteNpmTask;
@@ -25,7 +26,6 @@ import com.liferay.gradle.util.copy.RenameDependencyClosure;
 import java.io.File;
 
 import java.util.Collections;
-import java.util.HashSet;
 import java.util.Iterator;
 import java.util.Set;
 import java.util.concurrent.Callable;
@@ -35,8 +35,6 @@ import org.gradle.api.Plugin;
 import org.gradle.api.Project;
 import org.gradle.api.Task;
 import org.gradle.api.artifacts.Configuration;
-import org.gradle.api.artifacts.DependencySet;
-import org.gradle.api.artifacts.ProjectDependency;
 import org.gradle.api.file.FileCollection;
 import org.gradle.api.file.SourceDirectorySet;
 import org.gradle.api.plugins.BasePlugin;
@@ -124,44 +122,6 @@ public class JSTranspilerPlugin implements Plugin<Project> {
 		return downloadNodeModuleTask;
 	}
 
-	private Copy _addTaskExpandCompileDependency(
-		Project project, File file, File destinationDir, String taskNamePrefix,
-		RenameDependencyClosure renameDependencyClosure) {
-
-		String taskName = GradleUtil.getTaskName(taskNamePrefix, file);
-
-		Copy copy = GradleUtil.addTask(project, taskName, Copy.class);
-
-		copy.doFirst(
-			new Action<Task>() {
-
-				@Override
-				public void execute(Task task) {
-					Copy copy = (Copy)task;
-
-					Project project = copy.getProject();
-
-					project.delete(copy.getDestinationDir());
-				}
-
-			});
-
-		copy.from(project.zipTree(file));
-
-		String name = renameDependencyClosure.call(file.getName());
-
-		name = name.substring(0, name.length() - 4);
-
-		destinationDir = new File(destinationDir, name);
-
-		copy.setDescription(
-			"Expands " + file.getName() + " into " +
-				project.relativePath(destinationDir) + ".");
-		copy.setDestinationDir(destinationDir);
-
-		return copy;
-	}
-
 	private void _addTasksExpandSoyCompileDependencies(
 		TranspileJSTask transpileJSTask, Configuration configuration) {
 
@@ -170,11 +130,11 @@ public class JSTranspilerPlugin implements Plugin<Project> {
 		RenameDependencyClosure renameDependencyClosure =
 			new RenameDependencyClosure(project, configuration.getName());
 
-		Iterable<TaskDependency> taskDependencies = _getTaskDependencies(
-			configuration);
+		Iterable<TaskDependency> taskDependencies =
+			JSTranspilerPluginUtil.getTaskDependencies(configuration);
 
 		for (File file : configuration) {
-			Copy copy = _addTaskExpandCompileDependency(
+			Copy copy = JSTranspilerPluginUtil.addTaskExpandCompileDependency(
 				project, file, project.getBuildDir(),
 				"expandSoyCompileDependency", renameDependencyClosure);
 
@@ -335,22 +295,6 @@ public class JSTranspilerPlugin implements Plugin<Project> {
 		Iterator<File> iterator = srcDirs.iterator();
 
 		return iterator.next();
-	}
-
-	private Iterable<TaskDependency> _getTaskDependencies(
-		Configuration configuration) {
-
-		Set<TaskDependency> taskDependencies = new HashSet<>();
-
-		DependencySet dependencySet = configuration.getAllDependencies();
-
-		for (ProjectDependency projectDependency : dependencySet.withType(
-				ProjectDependency.class)) {
-
-			taskDependencies.add(projectDependency.getBuildDependencies());
-		}
-
-		return taskDependencies;
 	}
 
 	private static final String _METAL_CLI_VERSION = "1.3.1";
